fix registration components, service, servicecategory, service subcategory components and also fixed membership registration components

+++ app/Http/Controllers/ServiceProviderProfileCompletion/EducationQualificationController.php
@@ -4,14 +4,11 @@ namespace App\Http\Controllers\ServiceProviderProfileCompletion;
 
 use App\Contracts\Repositories\Registration\ServiceProviderRegistrationRepositoryInterface;
 use App\Http\Controllers\Controller;
-use App\Models\EducationQualification;
 use App\Models\ProviderQualification;
 use Illuminate\Http\Request;
 
 class EducationQualificationController extends Controller
 {
-
-
     public function store(Request $request, ServiceProviderRegistrationRepositoryInterface $repository)
     {
         $request->validate([
@@ -22,27 +19,31 @@ class EducationQualificationController extends Controller
             'attachment' => ['required', 'string', 'base64pdf']
         ]);
 
-        $data = $request->toArray();
+        if ($this->authorizeUser()) {
+            $data = $request->toArray();
 
-        $data+=['user_id' => auth()->user()->id];
+            $data += ['user_id' => auth()->user()->id];
 
-        $repository->createProviderQualification($data);
+            $repository->createProviderQualification($data);
+        }
 
         return redirect()->back()->with(['status' => 'Operation Complete successful']);
-
     }
 
-
+    public function authorizeUser(): bool
+    {
+        return auth()->user()->service_provider->is_submitted == false || auth()->user()->hasRole('verified-service-provider');
+    }
     public function destroy(
         ProviderQualification $qualification,
         ServiceProviderRegistrationRepositoryInterface $repository
-    )
-    {
+    ) {
+        if ($this->authorizeUser() &&
+            $qualification->provider_profile_id === auth()->user()->service_provider->id) {
+            $request['qualification_id'] = $qualification->id;
 
-        $request['qualification_id'] = $qualification->id;
-
-        $repository->deleteProviderQualification($request);
-
+            $repository->deleteProviderQualification($request);
+        }
         return redirect()->back()->with(['status' => 'Operation Complete successful']);
     }
 }
