+++ app/Http/Controllers/ServiceProviderProfileCompletion/ProviderFacilityController.php
@@ -6,12 +6,9 @@ use App\Contracts\Repositories\Registration\ServiceProviderRegistrationRepositor
 use App\Http\Controllers\Controller;
 use App\Models\ProviderFacility;
 use Illuminate\Http\Request;
-use phpDocumentor\Reflection\Utils;
 
 class ProviderFacilityController extends Controller
 {
-
-
     public function store(Request $request, ServiceProviderRegistrationRepositoryInterface $repository)
     {
         $this->validateFacility($request);
@@ -30,10 +27,8 @@ class ProviderFacilityController extends Controller
     public function update(
         Request $request,
         ProviderFacility $facility,
-        ServiceProviderRegistrationRepositoryInterface $repository)
-    {
-
-
+        ServiceProviderRegistrationRepositoryInterface $repository
+    ) {
         $this->validateFacility($request);
         $data = $request->toArray();
         $data+=['user_id' => auth()->user()->id];
@@ -42,12 +37,10 @@ class ProviderFacilityController extends Controller
         $repository->updateProviderFacility($data);
 
         return redirect()->back()->with(['status' => 'Operation Complete successful']);
-
     }
 
     public function CompanyInfoUpdate(Request $request, ServiceProviderRegistrationRepositoryInterface $repository)
     {
-
         $this->validateCompany($request);
 
         $data = $request->toArray();
@@ -57,22 +50,29 @@ class ProviderFacilityController extends Controller
         $repository->createOrUpdateProviderCompany($data);
 
         return redirect()->back()->with(['status' => 'Operation Complete successful']);
-
     }
 
+    public function authorizeUser(): bool
+    {
+        return auth()->user()->service_provider->is_submitted == false || auth()->user()->hasRole('verified-service-provider');
+    }
 
     public function destroy(
         ProviderFacility $facility,
-        ServiceProviderRegistrationRepositoryInterface $repository)
-    {
-        $request['facility_id'] = $facility->id;
+        ServiceProviderRegistrationRepositoryInterface $repository
+    ) {
+        if ($this->authorizeUser() &&
+            $facility->provider_company_id == auth()->user()->service_provider->provider_companies[0]->id) {
 
-        $repository->deleteProviderFacility($request);
+            $request['facility_id'] = $facility->id;
+            $repository->deleteProviderFacility($request);
+        }
 
         return redirect()->back()->with(['status' => 'Operation Complete successful']);
     }
 
-    private function validateFacility(Request $request){
+    private function validateFacility(Request $request)
+    {
         return $request->validate([
             'name' => ['required', 'max:255', 'string'],
             'trading_name' => ['required', 'max:255', 'string'],
@@ -89,10 +89,10 @@ class ProviderFacilityController extends Controller
             'description' => ['string', 'max:900'],
             'provider_sub_level_id' => ['required','numeric', 'exists:provider_sub_levels,id'],
         ]);
-
     }
 
-    private function validateCompany(Request $request){
+    private function validateCompany(Request $request)
+    {
         return $request->validate([
             'name' => ['required', 'max:255', 'string'],
             'trading_name' => ['required', 'max:255', 'string'],
@@ -108,6 +108,5 @@ class ProviderFacilityController extends Controller
             'registration_number' => ['required', 'max:255', 'string'],
             'description' => ['string', 'max:900'],
         ]);
-
     }
 }
